requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
openai==1.0.0 